#805 A1: harden and decompose agent CLI
