+++ packages/cli/package.json
@@ -7,6 +7,13 @@
   "bin": {
     "boring-ui": "./dist/index.js"
   },
+  "exports": {
+    "./server": {
+      "types": "./dist/server/cli.d.ts",
+      "import": "./dist/server/cli.js"
+    },
+    "./package.json": "./package.json"
+  },
   "files": [
     "dist/",
     "public/"
