+++ packages/cli/src/server/cli.ts
@@ -1,26 +1,20 @@
 import type { FastifyInstance } from "fastify"
 import type {
-  AgentHarnessFactory,
-  AuthoredAgentToolCatalog,
   BoringAgentRuntimePaths,
-  MaterializedAgentSourceV1,
   ProvisionWorkspaceRuntimeOptions,
   RuntimeModeAdapter,
   RuntimeModeId,
-  WorkspaceAgentDispatcherResolver,
   WorkspaceProvisioningAdapter,
   WorkspaceProvisioningResult,
 } from "@hachej/boring-agent/server"
 import { execSync } from "node:child_process"
-import { createHash } from "node:crypto"
-import {
-  existsSync,
-  readFileSync,
-} from "node:fs"
 import { createRequire } from "node:module"
 import { basename, isAbsolute, join, resolve } from "node:path"
 import { parseArgs } from "node:util"
 import { createLocalWorkspaceRegistry, type LocalWorkspace } from "./localWorkspaces.js"
+import { handleAgentCommand } from "./agentCommands.js"
+import type { RunCliAgentDevOptions } from "./agentCommandTypes.js"
+import { registerStatic } from "./staticAssets.js"
 import type {
   RuntimePluginDiagnosticsResponse,
   RuntimePluginHostSnapshot,
@@ -37,7 +31,7 @@ import {
   type RuntimeMode,
 } from "./modeApps.js"
 
-// Re-exported for existing importers (tests, embedding hosts).
+// Compatibility exports for existing internal/embedding importers.
 export {
   createBoringUiCliRuntimePlugin,
   createFolderModeApp,
@@ -46,21 +40,8 @@ export {
   resolveBoringUiPluginCliPackageRoot,
 } from "./modeApps.js"
 export { resolveBoringUiCliPackageRoot } from "./pluginDiscovery.js"
-
-export interface AgentDevTrustedToolCatalogAdapter {
-  resolveToolCatalog(input: {
-    directory: string
-    agentTypeId: string
-    declaredToolRefs: readonly string[]
-  }): AuthoredAgentToolCatalog | undefined | Promise<AuthoredAgentToolCatalog | undefined>
-}
-
-export interface RunCliAgentDevOptions {
-  trustedToolCatalogAdapter?: AgentDevTrustedToolCatalogAdapter
-  harnessFactory?: AgentHarnessFactory
-  runtimeModeAdapter?: RuntimeModeAdapter
-  provisionWorkspace?: boolean
-}
+export { registerStatic } from "./staticAssets.js"
+export type { AgentDevTrustedToolCatalogAdapter, RunCliAgentDevOptions } from "./agentCommandTypes.js"
 
 export interface RunCliOptions {
   argv?: string[]
@@ -79,52 +60,6 @@ function openBrowser(url: string) {
 
 // resolveBoringUiCliPackageRoot is imported from ./pluginDiscovery.js so
 // the discovery module can use the same helper without a circular import.
-
-
-function ensureFrontendBuilt(publicDir: string) {
-  if (existsSync(join(publicDir, "index.html"))) return
-  console.error("\nError: boring-ui frontend not found.")
-  console.error("Run `pnpm build:full` in packages/cli to build it first.\n")
-  process.exit(1)
-}
-
-export async function registerStatic(app: FastifyInstance, publicDir: string) {
-  ensureFrontendBuilt(publicDir)
-  // Compress responses (gzip/brotli) before serving static assets. The front
-  // bundle is multi-MB uncompressed; over a remote/tailscale link that raw
-  // transfer dominates first-load time. Compression cuts it ~3-4x. Registered
-  // before @fastify/static so its onSend hook wraps the file streams.
-  const { default: fastifyCompress } = await import("@fastify/compress")
-  await app.register(fastifyCompress, { global: true, encodings: ["br", "gzip"], threshold: 1024 })
-  const { default: fastifyStatic } = await import("@fastify/static")
-  await app.register(fastifyStatic, {
-    root: publicDir,
-    prefix: "/",
-    // @fastify/send writes its own Cache-Control after setHeaders runs, so
-    // disable it and set the header explicitly for both cases below.
-    cacheControl: false,
-    setHeaders(res, filePath) {
-      // Vite emits content-hashed filenames under /assets, so they can be
-      // cached forever — without this the multi-MB bundle is revalidated
-      // (or re-downloaded) on every workspace open. Everything else (notably
-      // index.html) keeps max-age=0 + etag so deploys are picked up
-      // immediately.
-      res.setHeader(
-        "cache-control",
-        /[\\/]assets[\\/]/.test(filePath)
-          ? "public, max-age=31536000, immutable"
-          : "public, max-age=0",
-      )
-    },
-  })
-
-  app.setNotFoundHandler(async (req, reply) => {
-    if (req.url.startsWith("/api/")) {
-      return reply.code(404).send({ error: "Not found" })
-    }
-    return reply.sendFile("index.html", publicDir)
-  })
-}
 
 const HELP_TEXT = [
   "Usage: boring-ui [workspace] [options]",
@@ -179,579 +114,6 @@ function truthyEnv(value: string | undefined): boolean {
   return value === "1" || value?.toLowerCase() === "true" || value?.toLowerCase() === "yes"
 }
 
-interface AgentCliErrorV1 {
-  schemaVersion: 1
-  ok: false
-  error: {
-    code: string
-    field?: string
-    message: string
-  }
-}
-
-interface AgentValidateSuccessV1 {
-  schemaVersion: 1
-  ok: true
-  agent: {
-    agentTypeId: string
-    version: string
-    label?: string
-    instructions: { present: true; byteLength: number }
-    refs: {
-      tools: string[]
-      capabilities: string[]
-      skills: string[]
-      mcpServers: string[]
-    }
-  }
-}
-
-class AgentValidateCliError extends Error {
-  readonly code: string
-  readonly field?: string
-
-  constructor(input: { code: string; field?: string; message: string }) {
-    super(input.message)
-    this.name = "AgentValidateCliError"
-    this.code = input.code
-    if (input.field !== undefined) this.field = input.field
-  }
-}
-
-const AGENT_TYPE_ID_RE = /^[a-z][a-z0-9-]{0,62}$/
-
-function copyRefs(refs: readonly string[] | undefined): string[] {
-  return refs === undefined ? [] : [...refs]
-}
-
-interface AgentValidateBundle {
-  definition: {
-    definitionId: string
-    version: string
-    label?: string
-    instructionsRef: string
-    capabilityRequirements?: readonly string[]
-    toolRefs?: readonly string[]
-    skillRefs?: readonly string[]
-    mcpServerRefs?: readonly string[]
-  }
-  assets: readonly { path: string; content: string }[]
-}
-
-interface AgentCommandDeps {
-  AgentDefinitionValidationError: new (...args: never[]) => Error & {
-    validationCode: string
-    field: string
-  }
-  AgentDirectoryCompilerError: new (...args: never[]) => Error & {
-    compilerCode: string
-    field: string
-  }
-  AuthoredAgentMaterializationError: new (...args: never[]) => Error & {
-    code: string
-    field?: string
-  }
-  compileAgentDirectory: (directory: string) => Promise<AgentValidateBundle>
-  materializeAgentDirectory: (input: {
-    directory: string
-    expectedAgentTypeId?: string
-    toolCatalog?: AuthoredAgentToolCatalog
-  }) => Promise<MaterializedAgentSourceV1>
-  createMaterializedAgentDevApp: typeof import("@hachej/boring-workspace/app/server")["createMaterializedAgentDevApp"]
-}
-
-async function loadAgentCommandDeps(): Promise<AgentCommandDeps> {
-  const [server, shared, workspaceAppServer] = await Promise.all([
-    import("@hachej/boring-agent/server"),
-    import("@hachej/boring-agent/shared"),
-    import("@hachej/boring-workspace/app/server"),
-  ])
-  return {
-    AgentDefinitionValidationError: shared.AgentDefinitionValidationError as AgentCommandDeps["AgentDefinitionValidationError"],
-    AgentDirectoryCompilerError: server.AgentDirectoryCompilerError as AgentCommandDeps["AgentDirectoryCompilerError"],
-    AuthoredAgentMaterializationError: server.AuthoredAgentMaterializationError as AgentCommandDeps["AuthoredAgentMaterializationError"],
-    compileAgentDirectory: server.compileAgentDirectory as AgentCommandDeps["compileAgentDirectory"],
-    materializeAgentDirectory: server.materializeAgentDirectory as AgentCommandDeps["materializeAgentDirectory"],
-    createMaterializedAgentDevApp: workspaceAppServer.createMaterializedAgentDevApp,
-  }
-}
-
-function createAgentValidateSuccess(bundle: AgentValidateBundle): AgentValidateSuccessV1 {
-  const { definition } = bundle
-  if (!AGENT_TYPE_ID_RE.test(definition.definitionId)) {
-    throw new AgentValidateCliError({
-      code: "AUTHORED_AGENT_ID_INVALID",
-      field: "definitionId",
-      message: "definitionId must match ^[a-z][a-z0-9-]{0,62}$",
-    })
-  }
-
-  const instructions = bundle.assets.find((asset) => asset.path === definition.instructionsRef)?.content
-  if (instructions === undefined) {
-    throw new AgentValidateCliError({
-      code: "INTERNAL_ERROR",
-      field: "instructionsRef",
-      message: "compiled agent instructions asset is missing",
-    })
-  }
-
-  return {
-    schemaVersion: 1,
-    ok: true,
-    agent: {
-      agentTypeId: definition.definitionId,
-      version: definition.version,
-      ...(definition.label === undefined ? {} : { label: definition.label }),
-      instructions: {
-        present: true,
-        byteLength: new TextEncoder().encode(instructions).byteLength,
-      },
-      refs: {
-        tools: copyRefs(definition.toolRefs),
-        capabilities: copyRefs(definition.capabilityRequirements),
-        skills: copyRefs(definition.skillRefs),
-        mcpServers: copyRefs(definition.mcpServerRefs),
-      },
-    },
-  }
-}
-
-function escapeTerminalUnsafeCharacter(value: string): string {
-  return value.replace(
-    /[\u007f-\u009f\u061c\u200e\u200f\u2028\u2029\u202a-\u202e\u2066-\u2069]/g,
-    (char) => `\\u${char.charCodeAt(0).toString(16).padStart(4, "0")}`,
-  )
-}
-
-function safeHumanValue(value: string): string {
-  return escapeTerminalUnsafeCharacter(value)
-}
-
-function safeHumanJsonValue(value: string): string {
-  return escapeTerminalUnsafeCharacter(JSON.stringify(value))
-}
-
-function refsLine(label: string, refs: readonly string[]): string {
-  return refs.length === 0 ? `    ${label}: 0` : `    ${label}: ${refs.length} (${refs.map(safeHumanValue).join(", ")})`
-}
-
-function formatAgentValidateHuman(payload: AgentValidateSuccessV1): string {
-  const lines = [
-    "Authored agent directory is valid.",
-    `  id: ${payload.agent.agentTypeId}`,
-    `  version: ${safeHumanValue(payload.agent.version)}`,
-  ]
-  if (payload.agent.label !== undefined) lines.push(`  label: ${safeHumanJsonValue(payload.agent.label)}`)
-  lines.push(
-    `  instructions: ${payload.agent.instructions.byteLength} bytes`,
-    "  declared refs:",
-    refsLine("tools", payload.agent.refs.tools),
-    refsLine("capabilities", payload.agent.refs.capabilities),
-    refsLine("skills", payload.agent.refs.skills),
-    refsLine("mcpServers", payload.agent.refs.mcpServers),
-  )
-  return lines.join("\n")
-}
-
-function toAgentCliError(error: unknown, deps?: AgentCommandDeps): AgentCliErrorV1 {
-  if (error instanceof AgentValidateCliError) {
-    return {
-      schemaVersion: 1,
-      ok: false,
-      error: {
-        code: error.code,
-        ...(error.field === undefined ? {} : { field: error.field }),
-        message: error.message,
-      },
-    }
-  }
-  if (deps !== undefined && error instanceof deps.AgentDirectoryCompilerError) {
-    return {
-      schemaVersion: 1,
-      ok: false,
-      error: {
-        code: error.compilerCode,
-        field: error.field,
-        message: error.message,
-      },
-    }
-  }
-  if (deps !== undefined && error instanceof deps.AgentDefinitionValidationError) {
-    return {
-      schemaVersion: 1,
-      ok: false,
-      error: {
-        code: error.validationCode,
-        field: error.field,
-        message: error.message,
-      },
-    }
-  }
-  if (deps !== undefined && error instanceof deps.AuthoredAgentMaterializationError) {
-    return {
-      schemaVersion: 1,
-      ok: false,
-      error: {
-        code: error.code,
-        ...(error.field === undefined ? {} : { field: error.field }),
-        message: error.message,
-      },
-    }
-  }
-  const stableError = error as { code?: unknown; field?: unknown; message?: unknown } | undefined
-  const stableCode = typeof stableError?.code === "string" ? stableError.code : ""
-  if (
-    typeof stableError?.message === "string" &&
-    (stableCode === "CONFIG_INVALID" || stableCode.startsWith("AUTHORED_AGENT_"))
-  ) {
-    return {
-      schemaVersion: 1,
-      ok: false,
-      error: {
-        code: stableCode,
-        ...(typeof stableError.field === "string" ? { field: stableError.field } : {}),
-        message: stableError.message,
-      },
-    }
-  }
-  return {
-    schemaVersion: 1,
-    ok: false,
-    error: {
-      code: "INTERNAL_ERROR",
-      message: "agent validation failed",
-    },
-  }
-}
-
-function unsupportedAgentValidateOption(token: string): AgentValidateCliError {
-  return new AgentValidateCliError({
-    code: "CONFIG_INVALID",
-    field: token.startsWith("--json=") ? "--json" : token.split("=", 1)[0],
-    message: "usage: boring-ui agent validate <dir> [--json]",
-  })
-}
-
-function parseAgentValidateArgv(argv: string[]): {
-  directory: string
-  json: boolean
-} {
-  const json = argv.includes("--json")
-  const agentIndex = argv.indexOf("agent")
-  if (agentIndex < 0) {
-    throw new AgentValidateCliError({
-      code: "CONFIG_INVALID",
-      field: "command",
-      message: "usage: boring-ui agent validate <dir>",
-    })
-  }
-
-  for (const token of argv.slice(0, agentIndex)) {
-    if (token === "--json") continue
-    if (token.startsWith("-")) throw unsupportedAgentValidateOption(token)
-  }
-
-  const tokens = argv.slice(agentIndex + 1).filter((token) => token !== "--json")
-  const subcommand = tokens[0]
-  if (subcommand !== "validate") {
-    throw new AgentValidateCliError({
-      code: "CONFIG_INVALID",
-      field: "command",
-      message: "usage: boring-ui agent validate <dir>",
-    })
-  }
-
-  let directory: string | undefined
-  for (const token of tokens.slice(1)) {
-    if (token.startsWith("-")) throw unsupportedAgentValidateOption(token)
-    if (directory !== undefined) {
-      throw new AgentValidateCliError({
-        code: "CONFIG_INVALID",
-        field: "arguments",
-        message: "usage: boring-ui agent validate <dir>",
-      })
-    }
-    directory = token
-  }
-  if (!directory) {
-    throw new AgentValidateCliError({
-      code: "CONFIG_INVALID",
-      field: "directory",
-      message: "usage: boring-ui agent validate <dir>",
-    })
-  }
-  return { directory, json }
-}
-
-interface AgentDevParsedArgv {
-  directory: string
-  prompt?: string
-  serve: boolean
-  allowDirect: boolean
-}
-
-function agentDevUsageError(field = "mode"): AgentValidateCliError {
-  return new AgentValidateCliError({
-    code: "AUTHORED_AGENT_DEV_USAGE_INVALID",
-    field,
-    message: "usage: boring-ui agent dev <dir> (--prompt <text> | --serve) [--allow-direct]",
-  })
-}
-
-function parseAgentDevArgv(argv: string[]): AgentDevParsedArgv {
-  const agentIndex = argv.indexOf("agent")
-  if (agentIndex < 0) throw agentDevUsageError("command")
-  const prefix = argv.slice(0, agentIndex)
-  for (let index = 0; index < prefix.length; index += 1) {
-    throw agentDevUsageError(prefix[index] || "arguments")
-  }
-
-  const tokens = argv.slice(agentIndex + 1)
-  if (tokens[0] !== "dev") throw agentDevUsageError("command")
-  let directory: string | undefined
-  let prompt: string | undefined
-  let serve = false
-  let allowDirect = false
-  for (let index = 1; index < tokens.length; index += 1) {
-    const token = tokens[index]
-    if (token === "--serve") {
-      if (serve) throw agentDevUsageError("--serve")
-      serve = true
-      continue
-    }
-    if (token === "--allow-direct") {
-      if (allowDirect) throw agentDevUsageError("--allow-direct")
-      allowDirect = true
-      continue
-    }
-    if (token === "--prompt") {
-      if (prompt !== undefined) throw agentDevUsageError("--prompt")
-      const value = tokens[index + 1]
-      if (value === undefined || value.trim().length === 0) throw agentDevUsageError("--prompt")
-      prompt = value
-      index += 1
-      continue
-    }
-    if (token?.startsWith("--prompt=")) {
-      if (prompt !== undefined) throw agentDevUsageError("--prompt")
-      const value = token.slice("--prompt=".length)
-      if (!value || value.trim().length === 0) throw agentDevUsageError("--prompt")
-      prompt = value
-      continue
-    }
-    if (token?.startsWith("-")) throw agentDevUsageError(token)
-    if (directory !== undefined) throw agentDevUsageError("arguments")
-    directory = token
-  }
-  if (!directory) throw agentDevUsageError("directory")
-  if ((prompt !== undefined && serve) || (prompt === undefined && !serve)) throw agentDevUsageError("mode")
-  return { directory, ...(prompt === undefined ? {} : { prompt }), serve, allowDirect }
-}
-
-function redactedWorkspaceId(workspaceId: string): string {
-  return `local:${createHash("sha256").update(workspaceId).digest("hex").slice(0, 10)}`
-}
-
-function terminalSafeId(value: string): string {
-  return safeHumanValue(value.replace(/[^A-Za-z0-9_.:-]/g, "_"))
-}
-
-function hasUnsupportedRefs(definition: AgentValidateBundle["definition"]): boolean {
-  return copyRefs(definition.capabilityRequirements).length > 0
-    || copyRefs(definition.skillRefs).length > 0
-    || copyRefs(definition.mcpServerRefs).length > 0
-}
-
-function sameRefs(a: readonly string[], b: readonly string[]): boolean {
-  return a.length === b.length && a.every((value, index) => value === b[index])
-}
-
-function stableAgentDevError(code: string, field: string | undefined, message: string): AgentValidateCliError {
-  return new AgentValidateCliError({
-    code,
-    ...(field === undefined ? {} : { field }),
-    message,
-  })
-}
-
-function assertAgentDevServeHostPolicy(): void {
-  const host = process.env.HOST ?? "127.0.0.1"
-  if (!isLoopbackHost(host)) throw agentDevUsageError("--host")
-}
-
-async function materializeAgentForDev(input: {
-  directory: string
-  deps: AgentCommandDeps
-  catalogAdapter?: AgentDevTrustedToolCatalogAdapter
-}): Promise<MaterializedAgentSourceV1> {
-  const bundle = await input.deps.compileAgentDirectory(input.directory)
-  const validated = createAgentValidateSuccess(bundle)
-  const declaredToolRefs = validated.agent.refs.tools
-  let toolCatalog: AuthoredAgentToolCatalog | undefined
-  if (declaredToolRefs.length > 0 && !hasUnsupportedRefs(bundle.definition)) {
-    toolCatalog = await input.catalogAdapter?.resolveToolCatalog({
-      directory: input.directory,
-      agentTypeId: validated.agent.agentTypeId,
-      declaredToolRefs,
-    })
-  }
-  const source = await input.deps.materializeAgentDirectory({
-    directory: input.directory,
-    expectedAgentTypeId: validated.agent.agentTypeId,
-    ...(toolCatalog === undefined ? {} : { toolCatalog }),
-  })
-  if (!sameRefs(source.declaredToolRefs, declaredToolRefs)) {
-    throw stableAgentDevError(
-      "AUTHORED_AGENT_REFERENCE_UNKNOWN",
-      "toolRefs",
-      "authored agent tool references changed during trusted catalog resolution",
-    )
-  }
-  return source
-}
-
-async function runAgentDevOneShot(input: {
-  resolver: WorkspaceAgentDispatcherResolver
-  workspaceId: string
-  sessionId: string
-  prompt: string
-}): Promise<void> {
-  const dispatcher = await input.resolver.resolve({ workspaceId: input.workspaceId, userId: "local-dev" })
-  let sawTerminalOk = false
-  for await (const event of dispatcher.send({ sessionId: input.sessionId, content: input.prompt })) {
-    // Intentionally consume but do not print model/tool output. The dev command
-    // reports only stable, redacted lifecycle identifiers.
-    const chunk = event.chunk
-    if (chunk.type === "error") {
-      throw stableAgentDevError("INTERNAL_ERROR", "turn", "authored agent dev turn failed")
-    }
-    if (chunk.type !== "agent-end") continue
-    if (chunk.willRetry === true) continue
-    if (chunk.status === "ok") {
-      sawTerminalOk = true
-      break
-    }
-    if (chunk.status === "aborted") {
-      throw stableAgentDevError("ABORTED", "turn", "authored agent dev turn aborted")
-    }
-    throw stableAgentDevError("INTERNAL_ERROR", "turn", "authored agent dev turn failed")
-  }
-  if (!sawTerminalOk) {
-    throw stableAgentDevError("INTERNAL_ERROR", "turn", "authored agent dev turn did not complete")
-  }
-}
-
-async function closeAppOnce(app: FastifyInstance): Promise<void> {
-  const state = app as FastifyInstance & { __boringAgentDevClosePromise?: Promise<void> }
-  state.__boringAgentDevClosePromise ??= app.close()
-  return await state.__boringAgentDevClosePromise
-}
-
-async function handleAgentDevCommand(
-  argv: string[],
-  options: RunCliOptions,
-  onDepsLoaded?: (deps: AgentCommandDeps) => void,
-): Promise<void> {
-  const parsed = parseAgentDevArgv(argv)
-  if (parsed.serve) assertAgentDevServeHostPolicy()
-  const deps = await loadAgentCommandDeps()
-  onDepsLoaded?.(deps)
-  let resolver: WorkspaceAgentDispatcherResolver | undefined
-  const source = await materializeAgentForDev({
-    directory: parsed.directory,
-    deps,
-    catalogAdapter: options.agentDev?.trustedToolCatalogAdapter,
-  })
-  const registry = createLocalWorkspaceRegistry()
-  const workspaceRoot = process.env.BORING_AGENT_WORKSPACE_ROOT ?? process.cwd()
-  const workspace = await registry.add(workspaceRoot, { name: basename(resolve(workspaceRoot)) || "workspace" })
-  const cliRuntimeMode: CliMode = parsed.allowDirect ? "local" : "local-sandbox"
-  const runtimeMode: RuntimeMode = parsed.allowDirect ? "direct" : "local"
-  const sessionId = `dev-${source.agentTypeId}`
-  const app = await deps.createMaterializedAgentDevApp({
-    source,
-    workspace: { root: workspace.path, sessionId: workspace.id },
-    runtime: {
-      mode: runtimeMode,
-      ...(options.agentDev?.runtimeModeAdapter ? { runtimeModeAdapter: options.agentDev.runtimeModeAdapter } : {}),
-      provisionWorkspace: options.agentDev?.provisionWorkspace ?? true,
-    },
-    harnessFactory: options.agentDev?.harnessFactory,
-    onWorkspaceAgentDispatcher: (value) => { resolver = value },
-  })
-  let serving = false
-  try {
-    if (!resolver) throw new Error("agent dev dispatcher was not initialized")
-    if (parsed.prompt !== undefined) {
-      await runAgentDevOneShot({ resolver, workspaceId: workspace.id, sessionId, prompt: parsed.prompt })
-      await closeAppOnce(app)
-      console.log("Authored agent dev one-shot completed.")
-      console.log(`  workspace   ${terminalSafeId(redactedWorkspaceId(workspace.id))}`)
-      console.log(`  agent type  ${terminalSafeId(source.agentTypeId)}`)
-      console.log(`  runtime     ${terminalSafeId(cliRuntimeMode)}`)
-      console.log(`  session     ${terminalSafeId(sessionId)}`)
-      return
-    }
-
-    const port = Number(process.env.PORT) || 5200
-    const host = "127.0.0.1"
-    await registerStatic(app, options.publicDir)
-    await app.listen({ port, host })
-    serving = true
-    console.log("Authored agent dev server ready.")
-    console.log(`  workspace   ${terminalSafeId(redactedWorkspaceId(workspace.id))}`)
-    console.log(`  agent type  ${terminalSafeId(source.agentTypeId)}`)
-    console.log(`  runtime     ${terminalSafeId(cliRuntimeMode)}`)
-    console.log(`  session     ${terminalSafeId(sessionId)}`)
-    console.log(`  url         http://${host}:${port}`)
-    let shutdownPromise: Promise<void> | undefined
-    const removeShutdownListeners = () => {
-      process.off("SIGINT", shutdown)
-      process.off("SIGTERM", shutdown)
-    }
-    const shutdown = () => {
-      shutdownPromise ??= (async () => {
-        try {
-          await closeAppOnce(app)
-          removeShutdownListeners()
-          process.exit(0)
-        } catch {
-          removeShutdownListeners()
-          console.error("INTERNAL_ERROR: \"authored agent dev server shutdown failed\"")
-          process.exit(1)
-        }
-      })()
-    }
-    process.on("SIGINT", shutdown)
-    process.on("SIGTERM", shutdown)
-  } finally {
-    if (parsed.prompt !== undefined || !serving) await closeAppOnce(app)
-  }
-}
-
-async function handleAgentCommand(argv: string[], options: RunCliOptions) {
-  let json = argv.includes("--json")
-  let deps: AgentCommandDeps | undefined
-  try {
-    const agentIndex = argv.indexOf("agent")
-    const subcommand = agentIndex >= 0 ? argv[agentIndex + 1] : undefined
-    if (subcommand === "dev") {
-      await handleAgentDevCommand(argv, options, (loadedDeps) => { deps = loadedDeps })
-      return
-    }
-
-    const parsed = parseAgentValidateArgv(argv)
-    json = parsed.json
-    deps = await loadAgentCommandDeps()
-    const payload = createAgentValidateSuccess(await deps.compileAgentDirectory(parsed.directory))
-    console.log(json ? JSON.stringify(payload) : formatAgentValidateHuman(payload))
-  } catch (error) {
-    const payload = toAgentCliError(error, deps)
-    const humanField = payload.error.field === undefined ? "" : ` ${safeHumanJsonValue(payload.error.field)}`
-    console.error(json ? JSON.stringify(payload) : `${payload.error.code}${humanField}: ${safeHumanJsonValue(payload.error.message)}`)
-    process.exitCode = payload.error.code === "AUTHORED_AGENT_DEV_USAGE_INVALID" ? 2 : 1
-  }
-}
 
 async function startFolderMode(opts: {
   folderArg?: string
