+++ packages/cli/tsup.config.ts
@@ -4,6 +4,13 @@ export default defineConfig({
   entry: [
     "src/index.ts",
     "src/server/cli.ts",
+    "src/server/agentCommands.ts",
+    "src/server/agentCommandDeps.ts",
+    "src/server/agentCommandSafe.ts",
+    "src/server/agentCommandTypes.ts",
+    "src/server/agentDevCommand.ts",
+    "src/server/agentValidateCommand.ts",
+    "src/server/staticAssets.ts",
     "src/server/localWorkspaces.ts",
     "src/server/modeApps.ts",
     "src/server/pluginDiscovery.ts",
@@ -13,6 +20,7 @@ export default defineConfig({
   format: ["esm"],
   target: "node20",
   bundle: false,
+  dts: true,
   clean: true,
   banner: { js: "#!/usr/bin/env node" },
 })
